requirements.txt
docker
python-pushover
TinyDB
